[#15] Feat : diaryHome 프래그먼트 통합
