--- app/src/main/res/layout/fragment_auto_created_today.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#FFFFFF"
-    tools:context=".AutoCreatedTodayFragment">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <TextView
-            android:id="@+id/diary_complete_tv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="38dp"
-            android:text="자동 일기 제작"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="20sp"
-            android:textStyle="bold"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"/>
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_refresh"
-            android:layout_marginStart="4dp"
-            android:layout_marginTop="4dp"
-            app:layout_constraintStart_toEndOf="@id/diary_complete_tv"
-            app:layout_constraintTop_toTopOf="@id/diary_complete_tv"
-            app:layout_constraintBottom_toBottomOf="@id/diary_complete_tv"/>
-
-        <ImageView
-            android:id="@+id/diary_complete_close_iv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_close"
-            android:layout_marginEnd="16dp"
-            app:layout_constraintTop_toTopOf="@id/diary_complete_tv"
-            app:layout_constraintBottom_toBottomOf="@id/diary_complete_tv"
-            app:layout_constraintEnd_toEndOf="parent"/>
-
-        <ImageView
-            android:id="@+id/diary_character_iv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_loading_star"
-            android:layout_marginTop="173dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/diary_complete_tv"/>
-
-        <TextView
-            android:id="@+id/diary_text_tv"
-            android:layout_width="292dp"
-            android:layout_height="164dp"
-            android:layout_marginTop="80dp"
-            android:text=""
-            android:textSize="14sp"
-            android:textColor="@color/black"
-            android:scrollbars="vertical"
-            android:paddingTop="15dp"
-            android:paddingBottom="49dp"
-            android:paddingHorizontal="16dp"
-            android:background="@drawable/ic_diary_box"
-            app:layout_constraintTop_toBottomOf="@id/diary_character_iv"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"/>
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_pen"
-            android:layout_marginEnd="15dp"
-            android:layout_marginBottom="15dp"
-            app:layout_constraintEnd_toEndOf="@id/diary_text_tv"
-            app:layout_constraintBottom_toBottomOf="@id/diary_text_tv"/>
-
-        <TextView
-            android:id="@+id/diary_imgAdd_tv"
-            android:layout_width="292dp"
-            android:layout_height="52dp"
-            android:text="사진 최대 4장 첨부해주세요."
-            android:fontFamily="@font/pretendard_regular"
-            android:layout_marginTop="24dp"
-            android:gravity="center"
-            android:textSize="16sp"
-            android:textStyle="bold"
-            android:background="@drawable/ic_question_box"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/diary_text_tv"
-            app:layout_constraintEnd_toEndOf="parent"
-            />
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_gallery"
-            android:layout_marginVertical="14dp"
-            android:layout_marginEnd="14dp"
-            app:layout_constraintEnd_toEndOf="@id/diary_imgAdd_tv"
-            app:layout_constraintBottom_toBottomOf="@id/diary_imgAdd_tv"/>
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</FrameLayout>
